pyproject.toml
[project]
name = "jl-hex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bokeh>=3.7.2",
    "holoviews>=1.20.2",
    "pillow>=11.1.0",
    "selenium>=4.31.0",
]
